fix: use the provided assertion to determine the rule type
If the provided assertion has run, it, and it alone, should be used to determine the rule type returned.

When a `true` value is returned, the rule type is returned as defined.

When a `false` value is returned, the rule type returned is the inverse of the one defined.

Fixes #2

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>