Cumulate all allow rules with assertions
This fixes the regression introduced in 624567fe376a70e0bfb5aa8217d5afa13b9d6e61 (2.7.2),
where rules with assertions were no longer cumulated, but instead the process was
aborted on first failed assertion and the privilege was incorrectly denied.

That regression directly contradicted that official documentation that states:

> a failed assertion results in the rule not applying, and other
> rules would be used to determine whether access is allowed or denied.

That regression prevented to fallback on more generic rules. And it also made it very brittle
to work with role inheritance since the order of inheritance declaration became critically important,
while adding/removing privileges anywhere in the inheritance chain could incorrectly have side effect
for other roles in an unexpected way.

So this patch is essentially a revert of the problematic commit. And the use-case at the origin of the
regression is now solved, and tested, differently, with proper usage of (pre-)existing behavior, as it
should have been from the beginning.

So instead of the incorrect:

```php
$acl = new Acl();

$assertAllow = new TestAsset\ChildBooleanAssertion(true);
$assertDeny  = new TestAsset\ChildBooleanAssertion(false);

$acl->addRole('staff');
$acl->addResource('base');
$acl->allow('staff', 'base', 'update', $assertAllow);

$acl->addResource('user', 'base');
$acl->allow('staff', 'user', 'update', $assertDeny);
```

It should be:

```php
$acl = new Acl();
$assertAllow = new MockAssertion(true);

$acl->addRole('staff');
$acl->addResource('base');
$acl->allow('staff', 'base', 'update', $assertAllow);

$acl->addResource('user', 'base');
$acl->deny('staff', 'user', 'update', $assertAllow);
```

Fixes #12

Signed-off-by: Adrien Crivelli <adrien.crivelli@gmail.com>